requirements.txt
fastapiuvicornstreamlitlangchainlangchain-communitylangchain-google-genailangchain-coreyoutube-transcript-apirequestspydantic